+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.17.1 under development
 
 - Enh #190: Open file in binary mode in `InputFile::fromLocalFile()` method.
+- New #191: Add `canManageBots` field to `User` type.
 
 ## 0.17 March 1, 2026